Simulate Data from Structural Equation Model

This is a thin wrapper for the simulateSEM()function in dagitty that works with tidied dagitty objects. It treats the input DAG as a structural equation model, generating random path coefficients and simulating corresponding data. See dagitty::simulateSEM() for details.

simulate_data(.tdy_dag, b.default = NULL, b.lower = -0.6,
  b.upper = 0.6, eps = 1, N = 500, standardized = TRUE)

Arguments

.tdy_dag
the input DAG, which can be a tidy_dagitty or dagitty object.

b.default
default path coefficient applied to arrows for which no coefficient is defined in the model syntax.

b.lower
lower bound for random path coefficients, applied if b.default = NULL.

b.upper
upper bound for path coefficients.

eps
residual variance (only meaningful if standardized=FALSE).

N
number of samples to generate.

standardized
whether a standardized output is desired (all variables have variance 1).

Value

a tblwith N values for each variable in .tdy_dag

Examples

dagify(y ~ z, x ~ z) %>%
  tidy_dagitty() %>%
  simulate_data()

#> # A tibble: 500 x 3
#>         x        y       z
#>     <dbl>    <dbl>   <dbl>
#>  1  0.610  0.0224  -0.741 
#>  2 -1.64  -0.589   -0.264 
#>  3  0.683  0.613    0.644 
#>  4  0.255 -1.44    -0.0738
#>  5 -0.277  0.951   -0.878 
#>  6 -0.776  0.536   -1.11  
#>  7 -2.02   0.00196  0.819 
#>  8 -1.67   1.34    -1.42  
#>  9 -0.843  1.31    -1.03  
#> 10 -1.01   1.42     0.112 
#> # … with 490 more rows