Simulate Data from Structural Equation Model

This is a thin wrapper for the simulateSEM()function in dagitty that works with tidied dagitty objects. It treats the input DAG as a structural equation model, generating random path coefficients and simulating corresponding data. See dagitty::simulateSEM() for details.

Usage

simulate_data(
  .tdy_dag,
  b.default = NULL,
  b.lower = -0.6,
  b.upper = 0.6,
  eps = 1,
  N = 500,
  standardized = TRUE
)

Arguments

.tdy_dag

the input DAG, which can be a tidy_dagitty or dagitty object.

b.default

default path coefficient applied to arrows for which no coefficient is defined in the model syntax.

b.lower

lower bound for random path coefficients, applied if b.default = NULL.

b.upper

upper bound for path coefficients.

eps

residual variance (only meaningful if standardized=FALSE).

N

number of samples to generate.

standardized

whether a standardized output is desired (all variables have variance 1).

Value

a tblwith N values for each variable in .tdy_dag

Examples

dagify(y ~ z, x ~ z) %>%
  tidy_dagitty() %>%
  simulate_data()
#> # A tibble: 500 × 3
#>         x       y       z
#>     <dbl>   <dbl>   <dbl>
#>  1  0.184 -1.74    0.939 
#>  2 -0.383  0.793   0.0930
#>  3  0.907  1.15    0.947 
#>  4 -1.55  -0.0210 -0.0470
#>  5 -0.889  1.96    2.72  
#>  6  0.135  0.902   0.0471
#>  7  1.58   1.19    0.111 
#>  8  0.334 -0.758   0.938 
#>  9  1.25   1.30    0.0955
#> 10  1.12   2.84    0.109 
#> # … with 490 more rows